implement abstract gui to simplify the implementation
