+++ Converter/src/main/java/de/jonas/informatik/converter/Converter.java
@@ -1,5 +1,7 @@
 package de.jonas.informatik.converter;
 
+import de.jonas.informatik.converter.gui.Gui;
+
 import java.util.HashMap;
 import java.util.Map;
